Use the size of args struct for deserializing
`ssmarshal` guarantees that the serialized size is <= the size of the
input type, but that's not necessarily the same as the sum of individual
struct members in the input type, since we're not packing them.

We _could_ evaluate whether packing makes sense, but for now, this
makes the server match the client (which already allocated
`size_of::<Func_ARGS>()` for serialization).