force REPLY type definitions into DWARF (#23)
The DWARF generated for types is load-bearing in that Humility
potentially needs them to be able to form arguments to Idol calls and
to make sense of the reply.  But if an Idol server is declared without
a consuming Idol client, any synthetic reply type won't be generated.
This work forces any synthetic reply type definition to be generated
by creating a meaningless static (that itself will be optimized away),
which has the side-effect of getting the type that we need in the
binary (but without changing the generated text or data).