Client support for non-u8 lease types.
The server-side runtime support has attempted to allow non-u8 lease
types since the beginning, but the client-side code generator is, uh,
more grown than designed. :-) This alters it to pass leased buffers
through zerocopy::AsBytes, allowing any type that implements the
appropriate zerocopy traits to be used.

Thanks to Github user @pokeylope for pointing this out and sending an
initial version of this commit.