reimplement codegen using `quote` (#40)
Currently, all of `idol`s code generation is implemented by "manually"
writing out Rust code as strings. The `quote` crate provides a macro for
quasi-quoting Rust code inside of Rust code, as a
`proc-_macro2::TokenStream`. This is generally the preferred way to do
code generation in much of the Rust ecosystem.

Some reasons we might want to use `quote` include:
- Integrating with `prettyplease`, a crate that does rustfmt-like
formatting as a library dep, so we no longer need to invoke an external
`rustfmt` as part of the build process in Hubris
- When using `quote`, we can interpolate things like types and
  identifiers as their parsed representation from the `syn` crate. This
  allows us to actually parse these fragments from the interface
  definition and validate that they're correct --- a `Name` can be
  represented by a `syn::Ident`, rather than a `String`, so we know that
  it's a valid Rust identifier, and parsing the RON will fail if it
  isn't. Similarly, we can parse types this way, allowing us to detect
  `!Sized` types much more accurately.
- Code quoted using `quote!` looks like Rust code, rather than a big
  string literal. Therefore, it can benefit from an editor's brace
  matching, syntax highlighting, indentation, etc.

This branch rewrites `idol`'s code generation using `quote`.

@cbiffle and I have both manually verified that Hubris builds with
its `idol` dependency pointed at this branch. We've also confirmed
that the generated code only differs in formatting and the addition
of `#[automatically_derived]` attributes for trait impls, which I added
in this branch.