fix compilation error in #44
Whoops, it looks like PR #44 broke the build. My bad --- I had
mistakenly assumed that we had a CI job that ensured that `idol` builds,
so I hadn't noticed that I introduced a compilation error. I'll go ahead
and set up CI builds for this repo, in a separate branch.