add a simple GitHub Actions CI workflow
Currently, this repo doesn't have any kind of CI set up to prevent us
from merging broken PRs. This has resulted in mistakes, at least on my
behalf (sorry). Therefore, this commit adds a quick CI setup for
ensuring that the code compiles and that it's been `rustfmt`ed
correctly.

We don't currently have any tests for the code in Idolatry, but if we
add some in the future, we'll want to add a job to run the tests on CI,
as well --- right now, it's just `cargo check` and `cargo fmt`.