make `idol` counter features opt-out
In general, we always want to generate IPC client counters, and only
disable them when building for targets that don't have the memory to
spare. In order to make it harder to accidentally forget to add counters
to a new task, this branch changes the feature flags from opting _in_ to
generating counters by default, to opting _out_ of generating counters
by default.