server: restore internal _args variable
This got changed to let _ = foo, which is great and correct.

However, it also causes current versions of rustc to become more
aggressive about not emitting the type in the DWARF, and this makes
Humility very sad.

I'm also patching Humility but, this fixes the code generator for older
versions of Humility.

This is the proximal cause of: https://github.com/oxidecomputer/hubris/issues/1777