remove zerocopy reexport
It turns out that the re-export of zerocopy doesn't actually help that
much, as crates embedding `idol`-generated code must still depend on
`zerocopy` explicitly. This is because the `zercopy_derive` macros
generate implementations for zerocopy's traits that reference them as
`zerocopy::<TRAIT>`, which breaks when there isn't a `zerocopy` crate in
scope when the generated code is compiled.