Update toolchain to `nightly-2025-07-20` (#65)
This branch updates `idol` to a recent toolchain, `nightly-2025-07-20`.
In order to generate code that compiles without warnings on this
toolchain, it was necessary to change a few things. 

`idol` used to put the `#[automatically_derived]` attribute on all
generated `impl` blocks. At some point between the previous ancient
toolchain and the one from a couple days ago, this attribute was changed
to only work on _trait impls_ and not _inherent impls_, and placing it
on inherent impls emits a warning. The server codegen only generates
trait impls, but the client code also generates an inherent impl for the
client struct (which is where most of the action happens), and it was
thus necessary  to remove the attribute from this. One of the things
`#[automatically_derived]` does is to suppress dead code and unused
variable warnings. Now that these warnings are no longer suppressed in
the client impl, some dead code and unused variables had to be removed
to make these warnings go away. In particular, it turns out that the
`<Name_of_interface>_<name_of_ipc>_REPLY` structs we used to generate
with zerocopy IPC clients were unused except by Humility, so I've added
`#[allow(dead_code)]` to them.

Additionally, I've changed the codegen for IPC counters to just use the
`counters!` macro to generate the static, instead of reimplementing it
in `idol` codegen. This way, we get whatever the `counters!` macro
generates, instead of having to keep it in sync. In particular, this
lets us benefit from the change to remove the `#[used]` attribute from
counters that was made in oxidecomputer/hubris#2167. Otherwise, it would
have been necessary to also remove that attribute from `idol`'s
reimplementation of the counters static generation.